requirements.txt
# Core Django & REST
django~=4.2.23
djangorestframework~=3.15.2
django-filter~=23.5
drf-yasg~=1.21.10

# Environment & CORS
django-environ~=0.11.2
django-cors-headers~=4.3.1

# Database
psycopg2~=2.9.9
django-redis~=5.4.0
django-cacheops~=7.0.0

# Authentication
cryptography==41.0.7
djangorestframework-simplejwt~=5.4.0
PyJWT==2.9.0

# Task Queue
celery~=5.4.0
redis~=5.0.3
django-celery-beat~=2.5.0


# GraphQL
graphene-django~=3.2.3
graphene-file-upload


# Channels & WebSockets
asgiref~=3.8.1
channels~=4.2.0
channels-redis~=4.2.0
daphne==4.1.2

# Others
httpagentparser~=1.9.5
django-mptt~=0.14.0
detoxify~=0.5.2
pillow

# Development & Testing
django-debug-toolbar~=4.3.0
pytest-django~=4.8.0
